refactor: collapse repository to mkiso
