+++ AGENTS.md
@@ -2,6 +2,8 @@
 
 This file applies to the entire repository.
 
+The project packages a single containerised script, `mkiso.py`. Keep the repository aligned with this layout when adding or updating files.
+
 ## Development workflow
 1. Format & lint all changes:
    ```
@@ -11,15 +13,10 @@ This file applies to the entire repository.
    ```
    pytest
    ```
-3. For container services with a `Containerfile`, update `spec.md`.
+3. Update `spec.md` whenever the CLI surface of `mkiso.py` changes.
 
 Avoid adding linter or formatter ignore comments (for example, `# noqa`, `# fmt: off`) unless absolutely necessary. When such comments are required, include an inline explanation of their necessity, justify why they are acceptable, and note this explicitly in the commit message.
 
-Scripts are containerised for cross‑platform use with Podman.
-
-When writing shell scripts, prefer POSIX-compliant `sh` and use other shells only when absolutely necessary.
-Each Podman script directory must include a `spec.md` written using modern Gauge conventions (`Scenario:` headings and `Given`/`When`/`Then` steps with parameter placeholders). Break steps into single, atomic actions and expectations to keep scenarios easily testable. For CLI specs, pass each flag or argument in its own step before the final run step. Specs should contain the minimal scenarios necessary for complete coverage of the script's features, including every command-line flag and argument. Use a single spec per platform-specific directory (e.g. `osx/spec.md`). If needed, keep script-specific terminology in a `glossary.md` alongside the spec.
-
-## Unit tests
-- Place tests in a `tests/` directory adjacent to the code.
-- Mock external commands and filesystem interactions; unit tests should exercise only the code under test.
+## Documentation and tests
+- Keep `spec.md` following modern Gauge conventions with `Scenario:` headings and `Given`/`When`/`Then` steps that map directly to the CLI actions.
+- Place tests in the `tests/` directory adjacent to `mkiso.py` and mock external commands or filesystem interactions so unit tests exercise only the code under test.
