+++ README.md
@@ -1,21 +1,46 @@
-# Scripts
+# mkiso
 
-Utility scripts for cross-platform workflows.
+mkiso is a containerised utility for packaging a directory tree into a boot-agnostic ISO image using `genisoimage`. The script is designed to run inside a Podman container so the host only needs a container runtime.
+
+## Repository layout
+
+- `Containerfile` – build definition for the runtime image.
+- `mkiso.py` – main entrypoint for the tool.
+- `spec.md` – Gauge specification describing expected behaviour.
+- `tests/` – unit tests for the script.
 
 ## Development
 
-Install and run the pre-commit hooks for changed files:
+Set up a virtual environment and install the tooling required for development:
 
 ```bash
-pre-commit install --install-hooks
-pre-commit run --files <files>
+python -m venv .venv
+. .venv/bin/activate
+pip install -U pip
+pip install pre-commit pytest
 ```
 
-Hooks cover formatting, linting, strict type checks with `mypy`, and Gauge spec validation.
-The Gauge CLI is installed automatically via pre-commit.
-
-Run tests:
+Run the automated checks before submitting changes:
 
 ```bash
+pre-commit run --all-files
 pytest
 ```
+
+Build and run the container image locally to exercise the workflow end-to-end:
+
+```bash
+podman build -t mkiso .
+podman run --rm -v "$PWD:/workspace" mkiso --help
+```
+
+## Release
+
+Cut a semantic version tag to publish a new container image to GitHub Container Registry:
+
+```bash
+git tag -a v1.2.3 -m "v1.2.3"
+git push origin v1.2.3
+```
+
+The release workflow builds multi-architecture images and pushes them to `ghcr.io/nashspence/mkiso` with the tag name, `latest`, and the Git commit SHA.
