--- containers/fittodisk/Containerfile
@@ -1,16 +1,0 @@
-# syntax=docker/dockerfile:1.7
-ARG TARGETPLATFORM
-ARG VERSION
-ARG VCS_REF
-ARG VCS_URL
-
-FROM --platform=$TARGETPLATFORM python:3.12-alpine
-
-LABEL org.opencontainers.image.source="${VCS_URL}" \
-      org.opencontainers.image.revision="${VCS_REF}" \
-      org.opencontainers.image.version="${VERSION}"
-
-RUN apk add --no-cache coreutils
-WORKDIR /app
-COPY script.py /app/fittodisk
-ENTRYPOINT ["python", "/app/fittodisk"]
