--- containers/guess_date/Containerfile
@@ -1,33 +1,0 @@
-# syntax=docker/dockerfile:1.7
-ARG TARGETPLATFORM
-ARG VERSION
-ARG VCS_REF
-ARG VCS_URL
-
-FROM --platform=$TARGETPLATFORM mwader/static-ffmpeg:latest AS ffmpeg
-
-FROM --platform=$TARGETPLATFORM python:3.12-alpine
-
-LABEL org.opencontainers.image.source="${VCS_URL}" \
-      org.opencontainers.image.revision="${VCS_REF}" \
-      org.opencontainers.image.version="${VERSION}"
-
-# Alpine packages (no apt). exiftool is Perl-based.
-RUN apk add --no-cache \
-      coreutils \
-      tzdata \
-      perl \
-      exiftool \
-      mediainfo
-
-# Bring in statically-built ffmpeg/ffprobe from mwader stage
-COPY --from=ffmpeg /ffmpeg /usr/local/bin/ffmpeg
-COPY --from=ffmpeg /ffprobe /usr/local/bin/ffprobe
-
-# Python deps
-RUN pip install --no-cache-dir python-dateutil
-
-WORKDIR /app
-COPY script.py /app/guess_date.py
-
-ENTRYPOINT ["python", "/app/guess_date.py"]
