--- containers/guess_date/release.yaml
@@ -1,5 +1,0 @@
-# single source of truth for this service
-image: docker.io/nashspence/scripts-guess-date
-version: v0.1.2
-labels:
-  org.opencontainers.image.title: guess_date
