--- containers/padimg/Containerfile
@@ -1,11 +1,0 @@
-FROM python:3.12-slim
-ARG VERSION
-ARG VCS_REF
-ARG VCS_URL
-LABEL org.opencontainers.image.source="${VCS_URL}" \
-      org.opencontainers.image.revision="${VCS_REF}" \
-      org.opencontainers.image.version="${VERSION}"
-RUN pip install --no-cache-dir pillow
-WORKDIR /app
-COPY script.py /app/padimg
-ENTRYPOINT ["python", "/app/padimg"]
