--- containers/stage/Containerfile
@@ -1,11 +1,0 @@
-FROM python:3.12-alpine
-ARG VERSION
-ARG VCS_REF
-ARG VCS_URL
-LABEL org.opencontainers.image.source="${VCS_URL}" \
-      org.opencontainers.image.revision="${VCS_REF}" \
-      org.opencontainers.image.version="${VERSION}"
-RUN apk add --no-cache coreutils
-WORKDIR /app
-COPY script.py /app/stage
-ENTRYPOINT ["python", "/app/stage"]
