--- containers/vcrunch/release.yaml
@@ -1,5 +1,0 @@
-# single source of truth for this service
-image: docker.io/nashspence/scripts-vcrunch
-version: v0.1.12
-labels:
-  org.opencontainers.image.title: vcrunch
