--- osx/bin/birthrname
@@ -1,56 +1,0 @@
-#!/bin/sh
-# Rename files in a directory by prefixing their creation time (APFS "birth time")
-# Format: YYYY-mm-ddTHH_MM_SS±ZZZZ!~filename[~N][.ext]
-
-set -eu
-[ -n "${DEBUG:-}" ] && set -x
-
-usage() {
-  printf 'Usage: %s <SRC_DIR>\n' "$0" >&2
-  exit 64
-}
-
-rename_files() {
-  dir=$1
-  find "$dir" -maxdepth 1 -type f ! -name '.*' -print |
-    while IFS= read -r filepath; do
-      filename=$(basename "$filepath")
-      creation_date=$(stat -f "%SB" -t "%Y-%m-%dT%H_%M_%S%z" "$filepath" 2>/dev/null) || {
-        printf 'stat failed: %s\n' "$filepath" >&2
-        exit 3
-      }
-      new_filename="${creation_date}!~${filename}"
-      new_filepath="$dir/$new_filename"
-      if [ "$filepath" != "$new_filepath" ]; then
-        if [ -e "$new_filepath" ]; then
-          stem=$filename
-          ext=
-          case $filename in
-            *.*) stem=${filename%.*}; ext=.${filename##*.} ;;
-          esac
-          i=1
-          while [ -e "$dir/${creation_date}!~${stem}~$i$ext" ]; do
-            i=$((i + 1))
-          done
-          new_filename="${creation_date}!~${stem}~$i$ext"
-          new_filepath="$dir/$new_filename"
-        fi
-        if ! mv "$filepath" "$new_filepath"; then
-          printf 'Rename failed: %s\n' "$filepath" >&2
-          exit 3
-        fi
-      fi
-    done
-}
-
-main() {
-  [ "$#" -eq 1 ] || usage
-  SRC_DIR=$1
-  [ -d "$SRC_DIR" ] || {
-    printf 'Not a directory: %s\n' "$SRC_DIR" >&2
-    exit 2
-  }
-  rename_files "$SRC_DIR"
-}
-
-main "$@"
