--- osx/bin/kc
@@ -1,20 +1,0 @@
-#!/bin/sh
-set -eu
-umask 077
-
-cmd="${1:?usage: $0 <add|get|delete> <name>}"
-service="com.nashspence.scripts.kc.${2:?usage: $0 <add|get|delete> <name>}"
-
-case "$cmd" in
-  add)
-    secret=$(cat)
-    [ -n "$secret" ] || security add-generic-password -s "$service" -a "$USER" -Uw && exit 0
-    security add-generic-password -s "$service" -a "$USER" -w "$secret" -U >/dev/null
-    ;;
-  get)
-    security find-generic-password -s "$service" -a "$USER" -w
-    ;;
-  delete)
-    security delete-generic-password -s "$service" -a "$USER" >/dev/null 2>&1 || true
-    ;;
-esac
