--- osx/bin/verifyburn
@@ -1,24 +1,0 @@
-#!/bin/sh
-# Usage: verify.sh image.dmg /dev/diskN
-
-set -eu
-
-IMG="$1"
-DISK="$2"
-
-# Attach image read-only, grab the whole-disk node
-IMGDEV=$(hdiutil attach -readonly -nomount "$IMG" | awk '/^\/dev\/disk[0-9]+/ {print $1; exit}')
-trap 'hdiutil detach "$IMGDEV" >/dev/null 2>&1 || true' EXIT
-
-# Use raw devices for speed
-IMGRAW=$(echo "$IMGDEV" | sed 's#/dev/disk#/dev/rdisk#')
-DISKRAW=$(echo "$DISK"   | sed 's#/dev/disk#/dev/rdisk#')
-
-# Silent byte-for-byte compare; exit 0 if identical
-if cmp -s "$IMGRAW" "$DISKRAW"; then
-  echo "OK: identical"
-  exit 0
-else
-  echo "FAIL: different"
-  exit 1
-fi
