--- osx/launch-agents/on-mount/.gitignore
@@ -1,2 +1,0 @@
-/data
-on-mount
