--- osx/launch-agents/on-mount/com.nashspence.scripts.on-mount.plist
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0"><dict>
-  <key>Label</key><string>com.nashspence.scripts.on-mount</string>
-
-  <key>ProgramArguments</key><array>
-    <string>%REPO_DIR%/osx/launch-agents/on-mount/on-mount</string>
-  </array>
-
-  <key>RunAtLoad</key><true/>
-  <key>KeepAlive</key><true/>
-  <key>LimitLoadToSessionType</key><string>Aqua</string>
-
-  <key>StandardOutPath</key>
-  <string>%REPO_DIR%/osx/launch-agents/on-mount/data/stdout.log</string>
-  <key>StandardErrorPath</key>
-  <string>%REPO_DIR%/osx/launch-agents/on-mount/data/stderr.log</string>
-</dict></plist>
