--- osx/launch-agents/podman-machine/.gitignore
@@ -1,1 +1,0 @@
-/data
