--- osx/launch-agents/podman-machine/com.nashspence.scripts.podman-machine.plist
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0"><dict>
-  <key>Label</key><string>com.nashspence.scripts.podman-machine</string>
-
-  <key>ProgramArguments</key>
-  <array>
-    <string>%REPO_DIR%/osx/launch-agents/podman-machine/Scripts - Podman Machine</string>
-  </array>
-
-  <key>Sockets</key>
-  <dict>
-    <key>Control</key>
-    <dict>
-      <key>SockPathName</key><string>/tmp/com.nashspence.scripts.podman-machine.%UID_NUM%.sock</string>
-      <key>SockFamily</key><string>Unix</string>
-      <key>SockType</key><string>stream</string>
-      <key>SockPathMode</key><integer>384</integer>
-    </dict>
-  </dict>
-
-  <key>inetdCompatibility</key>
-  <dict>
-    <key>Wait</key><false/>
-  </dict>
-
-  <key>EnvironmentVariables</key>
-  <dict>
-    <key>MACHINE</key><string>%MACHINE%</string>
-    <key>REPO_DIR</key><string>%REPO_DIR%</string>
-  </dict>
-
-  <key>KeepAlive</key><false/>
-  <key>ProcessType</key><string>Background</string>
-</dict></plist>
