--- scripts.code-workspace
@@ -1,12 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		}
-	],
-	"settings": {
-		"files.exclude": {
-			".*": true,
-		}
-	}
-}
